Move image and image question types to separate files.
