+++ src/image-question/components/app.tsx
@@ -2,29 +2,14 @@ import React from "react";
 import { JSONSchema6 } from "json-schema";
 import { BaseQuestionApp } from "../../shared/components/base-question-app";
 import { Runtime } from "./runtime";
-import { IAuthoringImageQuestionMetadata, IRuntimeImageQuestionMetadata } from "@concord-consortium/lara-interactive-api";
 import {
-  IAuthoredState as IDrawingToolAuthoredState, IInteractiveState as IDrawingToolInteractiveState
-} from "../../drawing-tool/components/types";
+  IAuthoredState,
+  IAuthoredStateBase,
+  IBaseInteractiveState,
+  IInteractiveState
+} from "./types";
 import { baseAuthoringProps as drawingToolBaseAuthoringProps } from "../../drawing-tool/components/app";
 import deepmerge from "deepmerge";
-
-export interface IAuthoredStateBase extends IAuthoringImageQuestionMetadata {
-  // IRuntimeImageQuestionMetadata adds:
-  // - answerType: "image_question"
-  // - answerPrompt?: string;
-  defaultAnswer?: string;
-}
-// Final IAuthoredState is combination of properties above and DrawingTool authored state (without questionType).
-export type IAuthoredState = IAuthoredStateBase & Omit<IDrawingToolAuthoredState, "questionType">;
-
-export interface IBaseInteractiveState extends IRuntimeImageQuestionMetadata {
-  // IRuntimeImageQuestionMetadata adds:
-  // - answerType: "image_question_interactive"
-  // - answerImageUrl?: string;
-  // - answerText?: string;
-}
-export type IInteractiveState = IBaseInteractiveState & Omit<IDrawingToolInteractiveState, "answerType">;
 
 const baseAuthoringProps = deepmerge(drawingToolBaseAuthoringProps, {
   schema: {
