+++ src/image-question/components/types.ts
@@ -1,0 +1,27 @@
+import {
+  IAuthoredState as IDrawingToolAuthoredState, 
+  IInteractiveState as IDrawingToolInteractiveState
+} from "../../drawing-tool/components/types";
+
+import { 
+  IAuthoringImageQuestionMetadata, 
+  IRuntimeImageQuestionMetadata 
+} from "@concord-consortium/lara-interactive-api";
+
+export interface IAuthoredStateBase extends IAuthoringImageQuestionMetadata {
+  // IRuntimeImageQuestionMetadata adds:
+  // - answerType: "image_question"
+  // - answerPrompt?: string;
+  defaultAnswer?: string;
+  customFeedback?: string;
+}
+// Final IAuthoredState is combination of properties above and DrawingTool authored state (without questionType).
+export type IAuthoredState = IAuthoredStateBase & Omit<IDrawingToolAuthoredState, "questionType">;
+
+export interface IBaseInteractiveState extends IRuntimeImageQuestionMetadata {
+  // IRuntimeImageQuestionMetadata adds:
+  // - answerType: "image_question_interactive"
+  // - answerImageUrl?: string;
+  // - answerText?: string;
+}
+export type IInteractiveState = IBaseInteractiveState & Omit<IDrawingToolInteractiveState, "answerType">;
