+++ src/image/components/runtime.tsx
@@ -1,5 +1,5 @@
 import React, { useRef } from "react";
-import { IAuthoredState } from "./app";
+import { IAuthoredState } from "./types";
 import { setSupportedFeatures, showModal } from "@concord-consortium/lara-interactive-api";
 import ReactDOMServer from "react-dom/server";
 import { log } from "@concord-consortium/lara-interactive-api";
